Move backfill parallel helper into facility_index_backfill utils
Keep post_deployment free of leftover temp wiring and document the shared helper for future deploy-time backfills.
